backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pycryptodome==3.21.0
python-dotenv==1.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
alembic==1.12.1 